backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-core==0.1.0
langchain-huggingface==0.0.1
langgraph==0.0.20
pymongo==4.6.0
reportlab==4.0.7
google-generativeai==0.3.2
